Write the shortDescription state variable once
`<barChart>` arrived with a third verbatim copy of it, 45 lines identical to
`<graph>`'s and `<image>`'s apart from the tag named in the accessibility
diagnostic and the noun in the public description. Both of those are now
arguments, and the definition lives in `utils/shortDescription` beside the
`utils/componentSize` and `utils/axisLabel` that `<graph>` and `<barChart>`
already share.

The schema regenerates unchanged, and the graph, image, prefigure and
diagnostics suites pass as they stood.
