+++ packages/doenetml-worker-javascript/src/components/Image.js
@@ -17,7 +17,7 @@ import {
     returnAnchorStateVariableDefinition,
 } from "../utils/graphical";
 import { returnListItemChildStateVariableDefinitions } from "../utils/listItemChild";
-import { codedDiagnostic } from "../utils/diagnostics";
+import { returnShortDescriptionStateVariableDefinition } from "../utils/shortDescription";
 
 export default class Image extends BlockComponent {
     constructor(args) {
@@ -279,53 +279,13 @@ export default class Image extends BlockComponent {
         let anchorDefinition = returnAnchorStateVariableDefinition();
         Object.assign(stateVariableDefinitions, anchorDefinition);
 
-        stateVariableDefinitions.shortDescription = {
-            description: "A short accessibility description of the image.",
-            forRenderer: true,
-            public: true,
-            shadowingInstructions: {
-                createComponentOfType: "text",
-            },
-            returnDependencies: () => ({
-                shortDescriptionChild: {
-                    dependencyType: "child",
-                    childGroups: ["shortDescriptions"],
-                    variableNames: ["text"],
-                },
-                decorative: {
-                    dependencyType: "stateVariable",
-                    variableName: "decorative",
-                },
+        Object.assign(
+            stateVariableDefinitions,
+            returnShortDescriptionStateVariableDefinition({
+                componentType: "image",
+                componentName: "image",
             }),
-            definition({ dependencyValues }) {
-                let shortDescription = "";
-                const diagnostics = [];
-                if (dependencyValues.shortDescriptionChild.length > 0) {
-                    const shortDescriptionChild =
-                        dependencyValues.shortDescriptionChild[
-                            dependencyValues.shortDescriptionChild.length - 1
-                        ];
-
-                    shortDescription =
-                        shortDescriptionChild.stateValues.text.trim();
-                }
-                if (shortDescription === "" && !dependencyValues.decorative) {
-                    diagnostics.push(
-                        codedDiagnostic({
-                            type: "accessibility",
-                            level: 1,
-                            code: "doenet-a0001",
-                            args: { component: "image" },
-                        }),
-                    );
-                }
-
-                return {
-                    setValue: { shortDescription },
-                    sendDiagnostics: diagnostics,
-                };
-            },
-        };
+        );
 
         stateVariableDefinitions.childIndicesToRender = {
             returnDependencies: () => ({
